saas-bi-agent/integrations/gemini_client.py
"""
Google Gemini API client wrapper.
Provides a unified interface for calling Gemini models with caching support.

TODO: Implement Gemini API integration
"""

from typing import Any, Dict, Optional, List
import google.generativeai as genai

from utils.config import get_config
from utils.logger import get_logger
from cache.cache_manager import CacheManager


class GeminiClient:
    """
    Wrapper for Google Gemini API with caching and error handling.

    Features:
    - Automatic prompt caching
    - Error handling and retries
    - Token counting
    - Safety settings management
    """

    def __init__(self, cache_manager: Optional[CacheManager] = None, api_key: Optional[str] = None):
        """
        Initialize Gemini client.

        Args:
            cache_manager: Optional cache manager for prompt caching
            api_key: Optional API key override
        """
        self.config = get_config()
        self.logger = get_logger(__name__)
        self.cache = cache_manager

        # Configure Gemini API
        # TODO: Implement API key configuration
        # genai.configure(api_key=api_key or os.getenv('GEMINI_API_KEY'))

        self.model_name = self.config.gemini.model
        self.logger.info(f"Initialized Gemini client with model: {self.model_name}")

    def generate(
        self,
        prompt: str,
        model: Optional[str] = None,
        temperature: Optional[float] = None,
        max_tokens: Optional[int] = None,
        use_cache: bool = True
    ) -> Dict[str, Any]:
        """
        Generate content using Gemini API.

        Args:
            prompt: The prompt to send
            model: Optional model override
            temperature: Optional temperature override
            max_tokens: Optional max tokens override
            use_cache: Whether to use prompt caching

        Returns:
            Dict with response, tokens, and metadata

        TODO: Implement actual Gemini API call
        """
        raise NotImplementedError("Gemini client not yet implemented")

    def count_tokens(self, text: str) -> int:
        """
        Count tokens in text.

        Args:
            text: Text to count

        Returns:
            Token count

        TODO: Implement token counting
        """
        # Rough estimation for now
        return int(len(text.split()) * 1.3)
